+++ .storybook/main.js
@@ -1,7 +1,12 @@
 module.exports = {
   stories: ['../packages/**/*.stories.@(tsx|mdx)'],
   addons: [
-    "@storybook/addon-essentials",
+    {
+      name: "@storybook/addon-essentials",
+      options: {
+        docs: false,
+      },
+    },
     "@storybook/addon-a11y",
     {
       name: "@storybook/addon-postcss",
@@ -15,7 +20,8 @@ module.exports = {
   typescript: {
     check: false,
     checkOptions: {},
-    reactDocgen: 'react-docgen-typescript',
+    // TODO: react-docgen-typescript when https://github.com/styleguidist/react-docgen-typescript/issues/356 gets resolved
+    reactDocgen: 'none',
     reactDocgenTypescriptOptions: {
       shouldExtractLiteralValuesFromEnum: true,
       propFilter: (prop) => (prop.parent ? !/node_modules/.test(prop.parent.fileName) : true),