+++ tw-plugins/core/src/classes/label.js
@@ -1,11 +1,11 @@
 const LabelBaseClasses = (theme) => ({
-  ".label": {
-    display: "flex",
-    alignItems: "center",
-    marginBottom: theme("spacing.2"),
-    fontSize: "1rem",
-    lineHeight: theme("spacing.6"),
-    color: theme("colors.gray.700"),
+  '.label': {
+    display: 'flex',
+    alignItems: 'center',
+    marginBottom: theme('spacing.2'),
+    fontSize: '1rem',
+    lineHeight: theme('spacing.6'),
+    color: theme('colors.gray.700'),
   },
 });