+++ tw-plugins/core/src/classes/pagination.js
@@ -1,12 +1,13 @@
 const PaginationBaseClasses = (theme) => ({
-  ".pagination": {
-    marginRight: theme("spacing.2"),
-    paddingLeft: theme("spacing.2"),
-    paddingRight: theme("spacing.2"),
-    boxShadow: theme("boxShadow.DEFAULT"),
-    color: theme("colors.blue.400"),
-    fontWeight: "500",
-    cursor: "pointer",
+  '.pagination': {
+    marginRight: theme('spacing.2'),
+    paddingLeft: theme('spacing.2'),
+    paddingRight: theme('spacing.2'),
+    borderRadius: theme('borderRadius.md'),
+    boxShadow: theme('boxShadow.DEFAULT'),
+    color: theme('colors.blue.400'),
+    fontWeight: '500',
+    cursor: 'pointer',
   },
 });