+++ tw-plugins/core/src/classes/skeleton.js
@@ -1,7 +1,7 @@
 const SliderBaseClasses = (theme) => ({
-  ".skeleton": {
-    backgroundColor: theme("colors.gray.200"),
-    borderRadius: theme("borderRadius.DEFAULT"),
+  '.skeleton': {
+    backgroundColor: theme('colors.gray.200'),
+    borderRadius: theme('borderRadius.DEFAULT'),
   },
 });