+++ tw-plugins/core/src/classes/tabs.js
@@ -1,8 +1,8 @@
 const TabsBaseClasses = (theme) => ({
-  ".tabs": {
-    display: "flex",
-    flexDirection: "column",
-    backgroundColor: theme("colors.gray.100"),
+  '.tabs': {
+    display: 'flex',
+    flexDirection: 'column',
+    backgroundColor: theme('colors.gray.100'),
   },
 });