+++ tw-plugins/core/src/classes/text-area.js
@@ -1,10 +1,10 @@
 const TextAreaBaseClasses = (theme) => ({
-  ".text-area": {
-    borderRadius: theme("borderRadius.md"),
-    borderWidth: theme("borderWidth.2"),
-    borderColor: theme("color.gray.300"),
-    padding: theme("spacing.2"),
-    maxWidth: theme("spacing.96"),
+  '.text-area': {
+    borderRadius: theme('borderRadius.md'),
+    borderWidth: theme('borderWidth.2'),
+    borderColor: theme('color.gray.300'),
+    padding: theme('spacing.2'),
+    maxWidth: theme('spacing.96'),
   },
 });