+++ tw-plugins/core/src/classes/text-input.js
@@ -3,10 +3,10 @@ const TextInputBaseClasses = (theme) => ({
     borderRadius: theme('borderRadius.md'),
     borderWidth: theme('borderWidth.2'),
     borderColor: theme('color.gray.200'),
-    padding: theme('spacing.2'),
+    padding: theme('input.py') || theme('spacing.2'),
     // height: theme('spacing.10'),
-    paddingLeft: theme('spacing.3'),
-    paddingRight: theme('spacing.3'),
+    paddingLeft: theme('input.px') || theme('spacing.3'),
+    paddingRight: theme('input.px') || theme('spacing.3'),
     // maxWidth: theme('spacing.96'),
   },
 });