+++ packages/core/CHANGELOG.md
@@ -1,5 +1,62 @@
 # @app-garage/core
 
+## 2.2.0
+
+### Minor Changes
+
+- [`41b9598`](https://github.com/electronic33/ag-ui-react/commit/41b959857d5143d241db516690b11d88c441f624) - Version dependencies
+
+## 2.1.0
+
+### Minor Changes
+
+- [`84d04fa`](https://github.com/electronic33/ag-ui-react/commit/84d04fa51dbf206cc4b2713796baeb2efbf54381) - Updated packages
+
+### Patch Changes
+
+- Updated dependencies [[`84d04fa`](https://github.com/electronic33/ag-ui-react/commit/84d04fa51dbf206cc4b2713796baeb2efbf54381), [`7392dc4`](https://github.com/electronic33/ag-ui-react/commit/7392dc47d3ccc528729c690ed036bb0aa41257ed)]:
+  - @app-garage/label@1.2.0
+  - @app-garage/accordion@1.2.0
+  - @app-garage/alert-dialog@1.1.0
+  - @app-garage/bottom-nav@1.1.0
+  - @app-garage/bread-crumbs@1.1.0
+  - @app-garage/button@1.2.0
+  - @app-garage/button-group@1.2.0
+  - @app-garage/calendar@1.1.0
+  - @app-garage/callout@1.1.0
+  - @app-garage/checkbox@1.1.0
+  - @app-garage/custom-select@1.1.0
+  - @app-garage/date-input@1.1.0
+  - @app-garage/drawer@1.1.0
+  - @app-garage/error-message@1.1.0
+  - @app-garage/focus-trap@1.1.0
+  - @app-garage/image-uploader@1.1.0
+  - @app-garage/link@1.1.0
+  - @app-garage/markdown@1.1.0
+  - @app-garage/menu-button@1.1.0
+  - @app-garage/menu@1.1.0
+  - @app-garage/modal@1.1.0
+  - @app-garage/multi-select@1.2.0
+  - @app-garage/numeric-input@1.1.0
+  - @app-garage/pagination@1.1.0
+  - @app-garage/popover@1.1.0
+  - @app-garage/progress@1.1.0
+  - @app-garage/radio-group@1.1.0
+  - @app-garage/rate@1.1.0
+  - @app-garage/scroll-to@1.1.0
+  - @app-garage/selectable-list@1.1.0
+  - @app-garage/skeleton@1.1.0
+  - @app-garage/slider@1.1.0
+  - @app-garage/spinner@1.1.0
+  - @app-garage/switch@1.1.0
+  - @app-garage/tabs@1.1.0
+  - @app-garage/text-area@1.1.0
+  - @app-garage/text-input@1.1.0
+  - @app-garage/toast@1.1.0
+  - @app-garage/toggle-button-group@1.1.0
+  - @app-garage/tooltip@1.1.0
+  - @app-garage/utils@1.1.0
+
 ## 2.0.1
 
 ### Patch Changes