+++ packages/multi-select/CHANGELOG.md
@@ -1,5 +1,35 @@
 # Change Log
 
+## 1.3.0
+
+### Minor Changes
+
+- [`f15d826`](https://github.com/electronic33/ag-ui-react/commit/f15d8268c75e59209d30a2509a76596130538333) - General improvements
+
+### Patch Changes
+
+- Updated dependencies [[`f15d826`](https://github.com/electronic33/ag-ui-react/commit/f15d8268c75e59209d30a2509a76596130538333), [`c15bb21`](https://github.com/electronic33/ag-ui-react/commit/c15bb21307eb1f0ead95009297c7e9e704ff1f1a)]:
+  - @app-garage/button@1.3.0
+  - @app-garage/custom-select@1.2.0
+
+## 1.2.0
+
+### Minor Changes
+
+- [`84d04fa`](https://github.com/electronic33/ag-ui-react/commit/84d04fa51dbf206cc4b2713796baeb2efbf54381) - Updated packages
+
+### Patch Changes
+
+- [`7392dc4`](https://github.com/electronic33/ag-ui-react/commit/7392dc47d3ccc528729c690ed036bb0aa41257ed) - downgrade yarn
+
+- Updated dependencies [[`84d04fa`](https://github.com/electronic33/ag-ui-react/commit/84d04fa51dbf206cc4b2713796baeb2efbf54381)]:
+  - @app-garage/button@1.2.0
+  - @app-garage/custom-select@1.1.0
+  - @app-garage/focus-trap@1.1.0
+  - @app-garage/spinner@1.1.0
+  - @app-garage/text-input@1.1.0
+  - @app-garage/utils@1.1.0
+
 ## 1.1.3
 
 ### Patch Changes