+++ packages/slider/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Change Log
 
+## 1.2.0
+
+### Minor Changes
+
+- [`c15bb21`](https://github.com/electronic33/ag-ui-react/commit/c15bb21307eb1f0ead95009297c7e9e704ff1f1a) - Fixe
+  zsh:1: command not found: q
+
+## 1.1.0
+
+### Minor Changes
+
+- [`84d04fa`](https://github.com/electronic33/ag-ui-react/commit/84d04fa51dbf206cc4b2713796baeb2efbf54381) - Updated packages
+
+### Patch Changes
+
+- Updated dependencies [[`84d04fa`](https://github.com/electronic33/ag-ui-react/commit/84d04fa51dbf206cc4b2713796baeb2efbf54381)]:
+  - @app-garage/modal@1.1.0
+
 ## 1.0.5
 
 ### Patch Changes